requirements.txt
google-api-python-client
google-cloud-storage
pydicom
Pillow
numpy